Added `tests` from `Part-2`
